# server/swagger.yml
openapi: "3.0.3"
info:
  title: SugarBox API by arairon
  version: 0.1.0
  description: A simple storage for SugarCube saves.
  license:
    name: GPL-3.0
    url: https://opensource.org/license/gpl-3-0
  contact:
    name: arai.icu
    url: https://arai.icu
    email: arai@arai.icu
servers:
  - url: https://sugarbox.arai.icu/api
  #- url: http://127.0.0.1:3000/api
paths:
...
  /user/session/{id}:
    delete:
      summary: Invalidate a session
      description: Invalidates a specific session for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The ID of the session to invalidate.
      responses:
        "200":
          description: Session invalidated successfully
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /user/sessions:
    get:
      summary: Get active sessions
      description: Retrieves all active sessions for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      responses:
        "200":
          description: Successfully retrieved active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    active:
                      type: boolean
                    tokens:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          active:
                            type: boolean
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /user/self:
    patch:
      summary: Update user information
      description: Updates the authenticated user's account information, such as username, email, or password.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPatchSchema"
      responses:
        "200":
          description: Successfully updated user information
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
        "400":
          description: Validation error or username/email already taken
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /user/quota:
    get:
      summary: Get storage quota
      description: Retrieves the authenticated user's storage quota and current usage.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      responses:
        "200":
          description: Successfully retrieved storage quota
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      usage:
                        type: integer
                      quota:
                        type: integer
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
...
  /sync/up:
    post:
      summary: Upload data for synchronization
      description: Uploads games, characters, and saves for synchronization.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncUpSchema"
      responses:
        "200":
          description: Successfully uploaded data for synchronization
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      errors:
                        type: array
                        items:
                          type: string
        "400":
          description: Validation error
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /sync/down:
    get:
      summary: Download data for synchronization
      description: Downloads games, characters, and saves updated since the cutoff point.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: query
          name: cutoffPoint
          required: true
          schema:
            type: string
            format: date-time
          description: The cutoff point for retrieving updated data.
        - in: query
          name: games
          required: false
          schema:
            type: boolean
          description: Whether to include games in the synchronization.
        - in: query
          name: chars
          required: false
          schema:
            type: boolean
          description: Whether to include characters in the synchronization.
        - in: query
          name: saves
          required: false
          schema:
            type: boolean
          description: Whether to include saves in the synchronization.
      responses:
        "200":
          description: Successfully downloaded data for synchronization
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      games:
                        type: array
                        items:
                          $ref: "#/components/schemas/GameSchema"
                      chars:
                        type: array
                        items:
                          $ref: "#/components/schemas/CharSchema"
                      saves:
                        type: array
                        items:
                          $ref: "#/components/schemas/SaveSchema"
        "400":
          description: Validation error
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
components:
  schemas:
...
    UserPatchSchema:
      type: object
      properties:
        displayname:
          type: string
          description: Display name of the user.
          minLength: 3
          maxLength: 64
        username:
          type: string
          description: Username of the user.
          minLength: 3
          maxLength: 32
        email:
          type: string
          description: Email address of the user.
          format: email
        password:
          type: string
          description: Password of the user.
          minLength: 3
          maxLength: 256
...
    GameSchema:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the game.
        uuid:
          type: string
          format: uuid
          description: UUID of the game.
        name:
          type: string
          maxLength: 100
          description: Name of the game.
        shortname:
          type: string
          maxLength: 100
          description: Short name of the game.
        description:
          type: string
          maxLength: 256
          description: Description of the game.
        ownerId:
          type: integer
          description: ID of the owner of the game.
        paths:
          type: string
          description: JSON string representing an array of game paths.
        archived:
          type: boolean
          description: Whether the game is archived.
        archivedAt:
          type: string
          format: date-time
          description: Timestamp when the game was archived.
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the game was last updated.
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the game was created.
...
    CharSchema:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the character.
        uuid:
          type: string
          format: uuid
          description: UUID of the character.
        name:
          type: string
          maxLength: 100
          description: Name of the character.
        ownerId:
          type: integer
          description: ID of the owner of the character.
        gameId:
          type: string
          format: uuid
          description: UUID of the game associated with the character.
        slots:
          type: string
          description: JSON string representing an array of UUIDs or empty strings.
        archived:
          type: boolean
          description: Whether the character is archived.
        archivedAt:
          type: string
          format: date-time
          description: Timestamp when the character was archived.
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the character was last updated.
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the character was created.
...
    SaveSchema:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the save.
        uuid:
          type: string
          format: uuid
          description: UUID of the save.
        name:
          type: string
          maxLength: 100
          description: Name of the save.
        description:
          type: string
          maxLength: 256
          description: Description of the save.
        gameVersion:
          type: string
          description: Version of the game associated with the save.
        gameId:
          type: string
          format: uuid
          description: UUID of the game associated with the save.
        charId:
          type: string
          format: uuid
          description: UUID of the character associated with the save.
        ownerId:
          type: integer
          description: ID of the owner of the save.
        data:
          type: string
          maxLength: 1000000
          description: Save data (up to 1MB).
        size:
          type: integer
          description: Size of the save data in bytes.
        hash:
          type: string
          maxLength: 256
          description: Hash of the save data.
        archived:
          type: boolean
          description: Whether the save is archived.
        archivedAt:
          type: string
          format: date-time
          description: Timestamp when the save was archived.
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the save was last updated.
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the save was created.
...
    SyncUpSchema:
      type: object
      properties:
        games:
          type: array
          items:
            type: object
          description: Array of games to upload.
        chars:
          type: array
          items:
            type: object
          description: Array of characters to upload.
        saves:
          type: array
          items:
            type: object
          description: Array of saves to upload.